Testing moving out one type to grafana/ui
