+++ public/app/types/series.ts
@@ -1,13 +1,6 @@
 import { Moment } from 'moment';
 import { PluginMeta } from './plugins';
 
-export enum LoadingState {
-  NotStarted = 'NotStarted',
-  Loading = 'Loading',
-  Done = 'Done',
-  Error = 'Error',
-}
-
 export interface RawTimeRange {
   from: Moment | string;
   to: Moment | string;
